+++ Source/CanvasElement.h
@@ -76,7 +76,7 @@ protected:
    virtual void DrawContents(bool clamp, bool wrapped, ofVec2f offset) = 0;
    void DrawElement(bool clamp, bool wrapped, ofVec2f offset);
    void AddElementUIControl(IUIControl* control);
-   void GetDragDestinationData(ofVec2f dragOffset, int& newRow, int& newCol, double & newOffset) const;
+   void GetDragDestinationData(ofVec2f dragOffset, int& newRow, int& newCol, double& newOffset) const;
    ofRectangle GetRectAtDestination(bool clamp, bool wrapped, ofVec2f dragOffset) const;
    double GetStart(int col, double offset) const;
    double GetEnd(int col, double offset, double length) const;
@@ -89,18 +89,18 @@ protected:
 class NoteCanvasElement : public CanvasElement
 {
 public:
-   NoteCanvasElement(Canvas* canvas, int col, int row, float offset, float length);
+   NoteCanvasElement(Canvas* canvas, int col, int row, double offset, double length);
    static CanvasElement* Create(Canvas* canvas, int col, int row) { return new NoteCanvasElement(canvas, col, row, 0, 1); }
-   void SetVelocity(float vel) { mVelocity = vel; }
-   float GetVelocity() const { return mVelocity; }
+   void SetVelocity(double vel) { mVelocity = vel; }
+   double GetVelocity() const { return mVelocity; }
    void SetVoiceIdx(int voiceIdx) { mVoiceIdx = voiceIdx; }
    int GetVoiceIdx() const { return mVoiceIdx; }
    ModulationChain* GetPitchBend() { return &mPitchBend; }
    ModulationChain* GetModWheel() { return &mModWheel; }
    ModulationChain* GetPressure() { return &mPressure; }
-   float GetPan() { return mPan; }
-   void UpdateModulation(float pos);
-   void WriteModulation(float pos, float pitchBend, float modWheel, float pressure, float pan);
+   double GetPan() { return mPan; }
+   void UpdateModulation(double pos);
+   void WriteModulation(double pos, double pitchBend, double modWheel, double pressure, double pan);
 
    CanvasElement* CreateDuplicate() const override;
 
@@ -120,7 +120,7 @@ private:
    ModulationChain mPitchBend{ ModulationParameters::kDefaultPitchBend };
    ModulationChain mModWheel{ ModulationParameters::kDefaultModWheel };
    ModulationChain mPressure{ ModulationParameters::kDefaultPressure };
-   float mPan{ 0 };
+   double mPan{ 0 };
    Curve mPitchBendCurve{ ModulationParameters::kDefaultPitchBend };
    Curve mModWheelCurve{ ModulationParameters::kDefaultModWheel };
    Curve mPressureCurve{ ModulationParameters::kDefaultPressure };
@@ -130,7 +130,7 @@ private:
 class SampleCanvasElement : public CanvasElement
 {
 public:
-   SampleCanvasElement(Canvas* canvas, int col, int row, float offset, float length);
+   SampleCanvasElement(Canvas* canvas, int col, int row, double offset, double length);
    ~SampleCanvasElement();
    static CanvasElement* Create(Canvas* canvas, int col, int row) { return new SampleCanvasElement(canvas, col, row, 0, 1); }
    void SetSample(Sample* sample);
@@ -162,14 +162,14 @@ private:
 class EventCanvasElement : public CanvasElement
 {
 public:
-   EventCanvasElement(Canvas* canvas, int col, int row, float offset);
+   EventCanvasElement(Canvas* canvas, int col, int row, double offset);
    ~EventCanvasElement();
    static CanvasElement* Create(Canvas* canvas, int col, int row) { return new EventCanvasElement(canvas, col, row, 0); }
 
    CanvasElement* CreateDuplicate() const override;
 
    void SetUIControl(IUIControl* control);
-   void SetValue(float value) { mValue = value; }
+   void SetValue(double value) { mValue = value; }
    void Trigger(double time);
    void TriggerEnd(double time);
